+++ packages/atomic/src/components/commerce/atomic-commerce-recommendation-interface/atomic-commerce-recommendation-interface.tsx
@@ -109,8 +109,10 @@ export class AtomicCommerceRecommendationInterface
 
   /**
    * The commerce interface language.
+   *
+   * Will default to the value set in the Headless engine context if not provided.
    */
-  @Prop({reflect: true}) public language = 'en';
+  @Prop({reflect: true, mutable: true}) public language?: string;
 
   /**
    * Whether to enable analytics.
@@ -126,12 +128,6 @@ export class AtomicCommerceRecommendationInterface
       this,
       'CoveoAtomic'
     );
-
-    if (!this.commonInterfaceHelper.engineIsCreated()) {
-      return;
-    }
-
-    this.contextController = buildContext(this.bindings.engine);
   }
 
   public connectedCallback() {
@@ -160,6 +156,10 @@ export class AtomicCommerceRecommendationInterface
   @Watch('language')
   public updateLanguage() {
     if (!this.commonInterfaceHelper.engineIsCreated(this.engine)) {
+      return;
+    }
+
+    if (!this.language) {
       return;
     }
 
@@ -233,6 +233,16 @@ export class AtomicCommerceRecommendationInterface
     };
   }
 
+  private initContext() {
+    this.contextController = buildContext(this.bindings.engine);
+  }
+
+  private initLanguage() {
+    if (!this.language) {
+      this.language = this.contextController.state.language;
+    }
+  }
+
   private initAriaLive() {
     if (
       Array.from(this.host.children).some(
@@ -246,6 +256,8 @@ export class AtomicCommerceRecommendationInterface
 
   private async internalInitialization(initEngine: () => void) {
     await this.commonInterfaceHelper.onInitialization(initEngine);
+    this.initContext();
+    this.initLanguage();
   }
 
   private addResourceBundle(