+++ packages/headless/src/controllers/commerce/core/sub-controller/headless-sub-controller.test.ts
@@ -6,7 +6,9 @@ import {
   MockedCommerceEngine,
   buildMockCommerceEngine,
 } from '../../../../test/mock-engine-v2';
+import {buildMockProduct} from '../../../../test/mock-product';
 import * as DidYouMean from '../../search/did-you-mean/headless-did-you-mean';
+import {SearchSummaryState} from '../../search/summary/headless-search-summary';
 import * as CoreBreadcrumbManager from '../breadcrumb-manager/headless-core-breadcrumb-manager';
 import * as CoreFacetGenerator from '../facets/generator/headless-commerce-facet-generator';
 import * as CorePagination from '../pagination/headless-core-commerce-pagination';
@@ -26,6 +28,13 @@ import {
 describe('sub-controllers', () => {
   let engine: MockedCommerceEngine;
   const mockResponseIdSelector = jest.fn();
+  const mockIsLoadingSelector = jest.fn();
+  const mockNumberOfProductsSelector = jest.fn();
+  const mockErrorSelector = jest.fn();
+  const mockPageSelector = jest.fn();
+  const mockPerPageSelector = jest.fn();
+  const mockTotalEntriesSelector = jest.fn();
+  const mockAugmentSummary = jest.fn();
   const mockFetchProductsActionCreator = jest.fn();
   const mockFetchMoreProductsActionCreator = jest.fn();
   const mockFacetResponseSelector = jest.fn();
@@ -54,6 +63,13 @@ describe('sub-controllers', () => {
     beforeEach(() => {
       subControllers = buildSearchSubControllers(engine, {
         responseIdSelector: mockResponseIdSelector,
+        isLoadingSelector: mockIsLoadingSelector,
+        numberOfProductsSelector: mockNumberOfProductsSelector,
+        errorSelector: mockErrorSelector,
+        pageSelector: mockPageSelector,
+        perPageSelector: mockPerPageSelector,
+        totalEntriesSelector: mockTotalEntriesSelector,
+        enrichSummary: mockAugmentSummary,
         fetchProductsActionCreator: mockFetchProductsActionCreator,
         fetchMoreProductsActionCreator: mockFetchMoreProductsActionCreator,
         facetResponseSelector: mockFacetResponseSelector,
@@ -93,11 +109,21 @@ describe('sub-controllers', () => {
   });
 
   describe('#buildSearchAndListingsSubControllers', () => {
-    let subControllers: SearchAndListingSubControllers<Parameters>;
+    let subControllers: SearchAndListingSubControllers<
+      Parameters,
+      SearchSummaryState
+    >;
 
     beforeEach(() => {
       subControllers = buildSearchAndListingsSubControllers(engine, {
         responseIdSelector: mockResponseIdSelector,
+        isLoadingSelector: mockIsLoadingSelector,
+        numberOfProductsSelector: mockNumberOfProductsSelector,
+        errorSelector: mockErrorSelector,
+        pageSelector: mockPageSelector,
+        perPageSelector: mockPerPageSelector,
+        totalEntriesSelector: mockTotalEntriesSelector,
+        enrichSummary: mockAugmentSummary,
         fetchProductsActionCreator: mockFetchProductsActionCreator,
         fetchMoreProductsActionCreator: mockFetchMoreProductsActionCreator,
         facetResponseSelector: mockFacetResponseSelector,
@@ -203,12 +229,19 @@ describe('sub-controllers', () => {
 
   describe('#buildBaseSubControllers', () => {
     const slotId = 'recommendations-slot-id';
-    let subControllers: BaseSolutionTypeSubControllers;
+    let subControllers: BaseSolutionTypeSubControllers<SearchSummaryState>;
 
     beforeEach(() => {
       subControllers = buildBaseSubControllers(engine, {
         slotId,
         responseIdSelector: mockResponseIdSelector,
+        isLoadingSelector: mockIsLoadingSelector,
+        numberOfProductsSelector: mockNumberOfProductsSelector,
+        errorSelector: mockErrorSelector,
+        pageSelector: mockPageSelector,
+        perPageSelector: mockPerPageSelector,
+        totalEntriesSelector: mockTotalEntriesSelector,
+        enrichSummary: mockAugmentSummary,
         fetchProductsActionCreator: mockFetchProductsActionCreator,
         fetchMoreProductsActionCreator: mockFetchMoreProductsActionCreator,
       });
@@ -222,12 +255,13 @@ describe('sub-controllers', () => {
 
       const props = {
         options: {
-          product: {
-            productId: '1',
-            name: 'Product name',
-            price: 17.99,
-          },
-          position: 1,
+          product: buildMockProduct({
+            ec_product_id: '1',
+            ec_name: 'Product name',
+            ec_promo_price: 15.99,
+            ec_price: 17.99,
+            position: 1,
+          }),
         },
       };