+++ packages/headless/src/features/facets/facet-search-set/generic/generic-facet-search-actions.ts
@@ -40,7 +40,10 @@ const getExecuteFacetSearchThunkPayloadCreator =
     ExecuteFacetSearchThunkArg,
     ExecuteFacetSearchThunkApiConfig
   > =>
-  async (facetId: string, {getState, extra: {apiClient, validatePayload}}) => {
+  async (
+    facetId: string,
+    {getState, extra: {apiClient, validatePayload, relay, navigatorContext}}
+  ) => {
     const state = getState();
     let req: SpecificFacetSearchRequest | CategoryFacetSearchRequest;
     validatePayload(facetId, requiredNonEmptyString);
@@ -48,12 +51,16 @@ const getExecuteFacetSearchThunkPayloadCreator =
       req = await buildSpecificFacetSearchRequest(
         facetId,
         state,
+        navigatorContext,
+        relay,
         isFieldSuggestionsRequest
       );
     } else {
       req = await buildCategoryFacetSearchRequest(
         facetId,
         state as StateNeededForCategoryFacetSearch,
+        navigatorContext,
+        relay,
         isFieldSuggestionsRequest
       );
     }