+++ packages/headless/src/features/product-listing/product-listing-slice.test.ts
@@ -1,19 +1,19 @@
 import {buildMockFacetResponse} from '../../test/mock-facet-response';
-import {buildFetchProductListingResponse} from '../../test/mock-product-listing';
+import {buildFetchProductListingResponse} from '../../test/mock-old-product-listing';
 import {buildMockProductRecommendation} from '../../test/mock-product-recommendation';
+import {
+  getProductListingInitialState,
+  OldProductListingState,
+} from './old-product-listing-state';
 import {
   fetchProductListing,
   setAdditionalFields,
   setProductListingUrl,
 } from './product-listing-actions';
 import {productListingReducer} from './product-listing-slice';
-import {
-  getProductListingInitialState,
-  ProductListingState,
-} from './product-listing-state';
 
 describe('product-listing-slice', () => {
-  let state: ProductListingState;
+  let state: OldProductListingState;
   beforeEach(() => {
     state = getProductListingInitialState();
   });