+++ packages/headless/src/integration-tests/insight-analytics-migration.test.ts
@@ -62,11 +62,7 @@ import {
   logInsightInterfaceChange,
   logInsightInterfaceLoad,
 } from '../features/insight-search/insight-search-analytics-actions';
-import {
-  pagerNext,
-  pagerNumber,
-  pagerPrevious,
-} from '../features/pagination/pagination-analytics-actions';
+import {browseResults} from '../features/pagination/pagination-analytics-actions';
 import {
   logPageNext,
   logPagePrevious,
@@ -82,7 +78,11 @@ import {
 } from '../features/static-filter-set/static-filter-set-actions';
 import {logInsightStaticFilterDeselect} from '../features/static-filter-set/static-filter-set-insight-analytics-actions';
 import {clearMicrotaskQueue} from '../test/unit-test-utils';
-import {assertNextEqualsLegacy} from './analytics-migration.test';
+import {
+  assertActionCause,
+  assertNextEqualsLegacy,
+  excludedBaseProperties,
+} from './analytics-migration.test';
 
 function getSampleInsightEngineConfiguration(): InsightEngineConfiguration {
   return {
@@ -442,40 +442,49 @@ describe('Analytics Search Migration', () => {
   it('analytics/pager/next', async () => {
     const action = fetchPage({
       legacy: logPageNext(),
-      next: pagerNext(),
+      next: browseResults(),
     });
 
     legacyInsightEngine.dispatch(action);
+    await clearMicrotaskQueue();
     nextInsightEngine.dispatch(action);
     await clearMicrotaskQueue();
 
-    assertNextEqualsLegacy(callSpy);
+    assertNextEqualsLegacy(callSpy, [...excludedBaseProperties, 'actionCause']);
+    assertActionCause(callSpy, 1, 'pagerNext');
+    assertActionCause(callSpy, 2, 'browseResults');
   });
 
   it('analytics/pager/previous', async () => {
     const action = fetchPage({
       legacy: logPagePrevious(),
-      next: pagerPrevious(),
+      next: browseResults(),
     });
 
     legacyInsightEngine.dispatch(action);
+    await clearMicrotaskQueue();
     nextInsightEngine.dispatch(action);
     await clearMicrotaskQueue();
 
-    assertNextEqualsLegacy(callSpy);
+    assertNextEqualsLegacy(callSpy, [...excludedBaseProperties, 'actionCause']);
+    assertActionCause(callSpy, 1, 'pagerPrevious');
+    assertActionCause(callSpy, 2, 'browseResults');
   });
 
   it('analytics/pager/number', async () => {
     const action = fetchPage({
       legacy: logPageNumber(),
-      next: pagerNumber(),
+      next: browseResults(),
     });
 
     legacyInsightEngine.dispatch(action);
+    await clearMicrotaskQueue();
     nextInsightEngine.dispatch(action);
     await clearMicrotaskQueue();
 
-    assertNextEqualsLegacy(callSpy);
+    assertNextEqualsLegacy(callSpy, [...excludedBaseProperties, 'actionCause']);
+    assertActionCause(callSpy, 1, 'pagerNumber');
+    assertActionCause(callSpy, 2, 'browseResults');
   });
 
   it('analytics/facet/showMore', async () => {